refactor(skuld): extract skuld to standalone top-level package
Move src/volundr/skuld/ to src/skuld/ — the broker is a standalone
runtime, not part of the Volundr API package. Extract LoggingConfig
to niuu.domain.logging for shared use. Update all imports, tests,
container entrypoints, and pyproject.toml.

Also create charts/skuld-planner/ Helm chart: a stripped-down copy
of charts/skuld/ without vscode-reh, devrunner, terminal debug, or
VS Code volume mounts. Lighter resource defaults for planning-only
sessions.
