+++ charts/skuld-planner/Chart.yaml
@@ -1,0 +1,18 @@
+apiVersion: v2
+name: skuld-planner
+description: "Skuld Planner — lightweight AI planning session broker (no IDE, no terminal)"
+type: application
+version: 0.0.0-tyr.150
+appVersion: "0.1.0"
+keywords:
+  - claude
+  - code
+  - session
+  - websocket
+  - planner
+home: https://github.com/niuulabs/volundr
+sources:
+  - https://github.com/niuulabs/volundr
+maintainers:
+  - name: Niuu Labs
+    url: https://github.com/niuulabs
