+++ charts/skuld-planner/templates/httproute.yaml
@@ -1,0 +1,36 @@
+{{- if .Values.gateway.enabled -}}
+# Gateway API HTTPRoute for session routing
+# Routes traffic from the shared Gateway to this session's services:
+#   /s/{session-id}/api       -> skuld (broker)
+#   /s/{session-id}/code      -> code-server
+#   /s/{session-id}/terminal  -> ttyd
+apiVersion: gateway.networking.k8s.io/v1
+kind: HTTPRoute
+metadata:
+  name: session-{{ .Values.session.id }}
+  labels:
+    {{- include "skuld.labels" . | nindent 4 }}
+    volundr.niuu.io/session: {{ .Values.session.id | quote }}
+    {{- with .Values.gateway.userId }}
+    volundr.niuu.io/user: {{ . | quote }}
+    {{- end }}
+spec:
+  parentRefs:
+    - name: {{ .Values.gateway.name }}
+      namespace: {{ .Values.gateway.namespace }}
+  rules:
+    # All session traffic — single prefix covers API, WebSocket, code-server, terminal
+    - matches:
+        - path:
+            type: PathPrefix
+            value: /s/{{ .Values.session.id }}
+      filters:
+        - type: URLRewrite
+          urlRewrite:
+            path:
+              type: ReplacePrefixMatch
+              replacePrefixMatch: /
+      backendRefs:
+        - name: {{ include "skuld.fullname" . }}
+          port: {{ .Values.service.port }}
+{{- end }}
