+++ charts/skuld-planner/templates/nginx-configmap.yaml
@@ -1,0 +1,98 @@
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  name: {{ include "skuld.fullname" . }}-nginx
+  labels:
+    {{- include "skuld.labels" . | nindent 4 }}
+data:
+  nginx.conf: |
+    worker_processes auto;
+    error_log /dev/stderr warn;
+    pid /tmp/nginx.pid;
+
+    events {
+        worker_connections 256;
+    }
+
+    http {
+        access_log /dev/stdout;
+
+        # Temp paths writable by non-root
+        client_body_temp_path /tmp/client_temp;
+        proxy_temp_path /tmp/proxy_temp;
+        fastcgi_temp_path /tmp/fastcgi_temp;
+        uwsgi_temp_path /tmp/uwsgi_temp;
+        scgi_temp_path /tmp/scgi_temp;
+
+        # Timeouts for long-running connections (WebSocket)
+        proxy_read_timeout 3600s;
+        proxy_send_timeout 3600s;
+        proxy_connect_timeout 10s;
+
+        # File upload support
+        client_max_body_size 100m;
+
+        # Upstream definitions - static backends
+        upstream broker {
+            server 127.0.0.1:8081;
+        }
+
+        server {
+            listen 8080;
+            server_name _;
+
+            # Broker: WebSocket session endpoint
+            location = {{ .Values.ingress.paths.session }} {
+                proxy_pass http://broker{{ .Values.ingress.paths.session }};
+                proxy_http_version 1.1;
+                proxy_set_header Upgrade $http_upgrade;
+                proxy_set_header Connection "upgrade";
+                proxy_set_header Host $host;
+                proxy_set_header X-Real-IP $remote_addr;
+                proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+                proxy_set_header X-Forwarded-Proto $scheme;
+            }
+            location {{ .Values.ingress.paths.session }}/ {
+                proxy_pass http://broker{{ .Values.ingress.paths.session }}/;
+                proxy_http_version 1.1;
+                proxy_set_header Upgrade $http_upgrade;
+                proxy_set_header Connection "upgrade";
+                proxy_set_header Host $host;
+                proxy_set_header X-Real-IP $remote_addr;
+                proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+                proxy_set_header X-Forwarded-Proto $scheme;
+            }
+
+            # Broker: REST API endpoints
+            location = /api {
+                return 301 /api/;
+            }
+            location /api/ {
+                proxy_pass http://broker/api/;
+                proxy_http_version 1.1;
+                proxy_set_header Host $host;
+                proxy_set_header X-Real-IP $remote_addr;
+                proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            }
+
+            # Broker: health and readiness (for probes)
+            location /health {
+                proxy_pass http://broker/health;
+            }
+
+            location /ready {
+                proxy_pass http://broker/ready;
+            }
+
+            # Default catch-all
+            location / {
+                proxy_pass http://broker/;
+                proxy_http_version 1.1;
+                proxy_set_header Upgrade $http_upgrade;
+                proxy_set_header Connection "upgrade";
+                proxy_set_header Host $host;
+                proxy_set_header X-Real-IP $remote_addr;
+                proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            }
+        }
+    }
