+++ charts/skuld-planner/templates/secret-manifest-configmap.yaml
@@ -1,0 +1,10 @@
+{{- if .Values.secretManifest }}
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  name: {{ include "skuld.fullname" . }}-secret-manifest
+  labels:
+    {{- include "skuld.labels" . | nindent 4 }}
+data:
+  manifest.json: {{ .Values.secretManifest | toJson | quote }}
+{{- end }}
