+++ charts/skuld-planner/templates/securitypolicy.yaml
@@ -1,0 +1,47 @@
+{{- if .Values.gateway.enabled -}}
+# Envoy SecurityPolicy for CORS and optional JWT validation.
+# CORS is required because the Volundr UI is served from a different
+# origin than the session gateway. JWT validation extracts the sub
+# claim into the x-auth-user-id header.
+apiVersion: gateway.envoyproxy.io/v1alpha1
+kind: SecurityPolicy
+metadata:
+  name: session-{{ .Values.session.id }}-security
+  labels:
+    {{- include "skuld.labels" . | nindent 4 }}
+    volundr.niuu.io/session: {{ .Values.session.id | quote }}
+spec:
+  targetRef:
+    group: gateway.networking.k8s.io
+    kind: HTTPRoute
+    name: session-{{ .Values.session.id }}
+  cors:
+    allowOrigins:
+      {{- toYaml .Values.gateway.cors.allowOrigins | nindent 6 }}
+    allowMethods:
+      {{- toYaml .Values.gateway.cors.allowMethods | nindent 6 }}
+    allowHeaders:
+      {{- toYaml .Values.gateway.cors.allowHeaders | nindent 6 }}
+    allowCredentials: {{ .Values.gateway.cors.allowCredentials }}
+    maxAge: "1h"
+  {{- if .Values.gateway.jwt.enabled }}
+  jwt:
+    providers:
+      - name: volundr-idp
+        issuer: {{ .Values.gateway.jwt.issuer | quote }}
+        audiences:
+          {{- toYaml .Values.gateway.jwt.audiences | nindent 10 }}
+        remoteJWKS:
+          uri: {{ .Values.gateway.jwt.jwksUri | quote }}
+          cacheDuration: 300s
+        extractFrom:
+          headers:
+            - name: Authorization
+              valuePrefix: "Bearer "
+          params:
+            - access_token
+        claimToHeaders:
+          - claim: sub
+            header: x-auth-user-id
+  {{- end }}
+{{- end }}
