+++ charts/skuld-planner/templates/skuld-configmap.yaml
@@ -1,0 +1,34 @@
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  name: {{ include "skuld.fullname" . }}-config
+  labels:
+    {{- include "skuld.labels" . | nindent 4 }}
+data:
+  config.yaml: |
+    # Skuld broker configuration (generated by Helm)
+
+    session:
+      id: {{ .Values.session.id | quote }}
+      name: {{ .Values.session.name | quote }}
+      model: {{ .Values.session.model | quote }}
+      {{- if .Values.session.systemPrompt }}
+      system_prompt: {{ .Values.session.systemPrompt | quote }}
+      {{- end }}
+      {{- if .Values.session.initialPrompt }}
+      initial_prompt: {{ .Values.session.initialPrompt | quote }}
+      {{- end }}
+
+    transport: {{ .Values.broker.transport | default "sdk" | quote }}
+    skip_permissions: {{ .Values.broker.skipPermissions | default true }}
+    agent_teams: {{ .Values.broker.agentTeams | default false }}
+    cli_type: {{ .Values.broker.cliType | default "claude" | quote }}
+    host: "0.0.0.0"
+    port: 8081
+    persistence_mount_path: {{ .Values.persistence.mountPath | default "/volundr/sessions" | quote }}
+
+    {{- if .Values.volundr.apiUrl }}
+    volundr_api_url: {{ .Values.volundr.apiUrl | quote }}
+    {{- end }}
+    service_user_id: {{ .Values.volundr.serviceUserId | default "skuld-broker" | quote }}
+    service_tenant_id: {{ .Values.volundr.serviceTenantId | default "default" | quote }}
