+++ pyproject.toml
@@ -49,7 +49,7 @@ dev = [
 
 [project.scripts]
 volundr = "volundr.main:main"
-skuld = "volundr.skuld.broker:main"
+skuld = "skuld.broker:main"
 bifrost = "volundr.bifrost.__main__:main"
 tyr = "tyr.main:main"
 
@@ -64,7 +64,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/niuu", "src/volundr", "src/tyr"]
+packages = ["src/niuu", "src/volundr", "src/tyr", "src/skuld"]
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
