+++ src/skuld/__main__.py
@@ -1,6 +1,6 @@
 """Skuld broker entrypoint."""
 
-from volundr.skuld.broker import main
+from skuld.broker import main
 
 if __name__ == "__main__":
     main()
