+++ tests/test_skuld/test_broker.py
@@ -8,15 +8,15 @@ import pytest
 from fastapi import WebSocketDisconnect
 from fastapi.testclient import TestClient
 
-from volundr.skuld.broker import (
+from skuld.broker import (
     Broker,
     _log_buffer,
     _TokenRedactFilter,
     app,
     broker,
 )
-from volundr.skuld.config import SkuldSettings
-from volundr.skuld.transport import (
+from skuld.config import SkuldSettings
+from skuld.transport import (
     CodexSubprocessTransport,
     SdkWebSocketTransport,
     SubprocessTransport,
@@ -652,7 +652,7 @@ class TestSessionArtifacts:
     """Tests for SessionArtifacts accumulator."""
 
     def test_record_tool_use_extracts_file_paths(self):
-        from volundr.skuld.broker import SessionArtifacts
+        from skuld.broker import SessionArtifacts
 
         artifacts = SessionArtifacts()
         data = {
@@ -675,7 +675,7 @@ class TestSessionArtifacts:
         assert artifacts.files_changed == ["/src/main.py", "/tests/test_main.py"]
 
     def test_record_tool_use_deduplicates(self):
-        from volundr.skuld.broker import SessionArtifacts
+        from skuld.broker import SessionArtifacts
 
         artifacts = SessionArtifacts()
         data = {"content": [{"type": "tool_use", "input": {"file_path": "/src/main.py"}}]}
@@ -684,7 +684,7 @@ class TestSessionArtifacts:
         assert artifacts.files_changed == ["/src/main.py"]
 
     def test_record_tool_use_empty_content(self):
-        from volundr.skuld.broker import SessionArtifacts
+        from skuld.broker import SessionArtifacts
 
         artifacts = SessionArtifacts()
         artifacts.record_tool_use({"content": []})
@@ -692,7 +692,7 @@ class TestSessionArtifacts:
         assert artifacts.files_changed == []
 
     def test_record_result_increments_turns(self):
-        from volundr.skuld.broker import SessionArtifacts
+        from skuld.broker import SessionArtifacts
 
         artifacts = SessionArtifacts()
         assert artifacts.turn_count == 0
@@ -701,7 +701,7 @@ class TestSessionArtifacts:
         assert artifacts.turn_count == 2
 
     def test_duration_seconds(self):
-        from volundr.skuld.broker import SessionArtifacts
+        from skuld.broker import SessionArtifacts
 
         artifacts = SessionArtifacts()
         # duration should be >= 0
@@ -1271,7 +1271,7 @@ class TestRecordToolUseReturnsEvents:
     """Tests for record_tool_use returning timeline-reportable events."""
 
     def test_returns_file_event_for_edit(self):
-        from volundr.skuld.broker import SessionArtifacts
+        from skuld.broker import SessionArtifacts
 
         artifacts = SessionArtifacts()
         data = {
@@ -1290,7 +1290,7 @@ class TestRecordToolUseReturnsEvents:
         assert events[0]["action"] == "modified"
 
     def test_returns_file_event_for_write(self):
-        from volundr.skuld.broker import SessionArtifacts
+        from skuld.broker import SessionArtifacts
 
         artifacts = SessionArtifacts()
         data = {
@@ -1308,7 +1308,7 @@ class TestRecordToolUseReturnsEvents:
         assert events[0]["action"] == "created"
 
     def test_returns_terminal_event_for_bash(self):
-        from volundr.skuld.broker import SessionArtifacts
+        from skuld.broker import SessionArtifacts
 
         artifacts = SessionArtifacts()
         data = {
@@ -1326,7 +1326,7 @@ class TestRecordToolUseReturnsEvents:
         assert "pytest" in events[0]["label"]
 
     def test_returns_git_event_for_git_commit(self):
-        from volundr.skuld.broker import SessionArtifacts
+        from skuld.broker import SessionArtifacts
 
         artifacts = SessionArtifacts()
         data = {
@@ -1344,7 +1344,7 @@ class TestRecordToolUseReturnsEvents:
         assert "git commit" in events[0]["label"]
 
     def test_returns_git_event_for_chained_commit(self):
-        from volundr.skuld.broker import SessionArtifacts
+        from skuld.broker import SessionArtifacts
 
         artifacts = SessionArtifacts()
         data = {
@@ -1361,7 +1361,7 @@ class TestRecordToolUseReturnsEvents:
         assert events[0]["type"] == "git"
 
     def test_returns_multiple_events(self):
-        from volundr.skuld.broker import SessionArtifacts
+        from skuld.broker import SessionArtifacts
 
         artifacts = SessionArtifacts()
         data = {
@@ -1384,7 +1384,7 @@ class TestRecordToolUseReturnsEvents:
         assert events[1]["type"] == "terminal"
 
     def test_returns_empty_for_unknown_tools(self):
-        from volundr.skuld.broker import SessionArtifacts
+        from skuld.broker import SessionArtifacts
 
         artifacts = SessionArtifacts()
         data = {
@@ -1402,7 +1402,7 @@ class TestRecordToolUseReturnsEvents:
         assert artifacts.files_changed == ["/src/a.py"]
 
     def test_returns_empty_for_no_content(self):
-        from volundr.skuld.broker import SessionArtifacts
+        from skuld.broker import SessionArtifacts
 
         artifacts = SessionArtifacts()
         events = artifacts.record_tool_use({})
@@ -1413,34 +1413,34 @@ class TestIsGitCommit:
     """Tests for _is_git_commit helper."""
 
     def test_simple_git_commit(self):
-        from volundr.skuld.broker import _is_git_commit
+        from skuld.broker import _is_git_commit
 
         assert _is_git_commit('git commit -m "msg"') is True
 
     def test_git_commit_with_flags(self):
-        from volundr.skuld.broker import _is_git_commit
+        from skuld.broker import _is_git_commit
 
         assert _is_git_commit("git commit --amend --no-edit") is True
 
     def test_chained_git_add_and_commit(self):
-        from volundr.skuld.broker import _is_git_commit
+        from skuld.broker import _is_git_commit
 
         assert _is_git_commit('git add . && git commit -m "feat"') is True
 
     def test_git_config_prefix(self):
-        from volundr.skuld.broker import _is_git_commit
+        from skuld.broker import _is_git_commit
 
         assert _is_git_commit('git -c user.name="x" commit -m "y"') is True
 
     def test_not_git_commit(self):
-        from volundr.skuld.broker import _is_git_commit
+        from skuld.broker import _is_git_commit
 
         assert _is_git_commit("git status") is False
         assert _is_git_commit("git push origin main") is False
         assert _is_git_commit("python -m pytest") is False
 
     def test_empty_string(self):
-        from volundr.skuld.broker import _is_git_commit
+        from skuld.broker import _is_git_commit
 
         assert _is_git_commit("") is False
 
@@ -1850,7 +1850,7 @@ class TestTokenRedactFilter:
         """Lifespan attaches redact filter to uvicorn loggers."""
         import asyncio
 
-        from volundr.skuld.broker import lifespan
+        from skuld.broker import lifespan
 
         async def check():
             for name in ("uvicorn", "uvicorn.access", "uvicorn.error"):
