+++ tests/test_skuld/test_file_listing.py
@@ -3,7 +3,7 @@
 import pytest
 from fastapi.testclient import TestClient
 
-from volundr.skuld.broker import app, broker
+from skuld.broker import app, broker
 
 
 class TestFileListingEndpoint:
